psy-bot-api/requirements.txt
Flask
flask-cors
openai
python-dotenv
peewee
psycopg2-binary
bcrypt
PyJWT
requests
